The explodeFireworks() method will be triggered when the user wants to set off their selected fireworks, so it loops through the fireworks array (backwards), picks out any selected ones, then calls explode() on it. Since the fireworks array stores the firework container node, the firework image is read out of its children array.
A method called motionBegan() has been written to detect when the device was shaken.

In addition, score label that updates as the player’s score changes has been added. Also, the game now ends after 10 launches. After 10 launches of fireworks, the gameOver() method is called, which calls the invalidate() method of Timer to stop it from repeating. In addition, the waitForDuration and removeFromParent actions have been used in a sequence to make sure explosion particle emitters are removed from the game scene when they are finished within the new explode(firework: SKNode) method (the method creates an explosion where the firework was).